anomaly-detection/requirements.txt
paho-mqtt
scikit-learn
influxdb-client
pandas
numpy